refactor: wmg pipeline cube creation (#2802)
* sp

* lint

* remove unused var

* refactor cell count cube

* lint, add logs

* lint again

* cleanup

* lint

* fmt

* cleanup

* add comments to functions

* lint

* more cleanup

* cleanup

* lint
