+++ backend/wmg/data/schemas/cube_schema.py
@@ -10,6 +10,11 @@ cube_indexed_dims = [
     "organism_ontology_term_id",
 ]
 
+cube_indexed_dims_no_gene_ontology = [
+    "tissue_ontology_term_id",
+    "organism_ontology_term_id",
+]
+
 # These are the queryable cube dimensions that will be modeled as
 # TileDB `Attrs` (i.e. (non-indexed") and thus will require
 # client-side filtering, which may result in less efficient querying.
