+++ backend/wmg/data/utils.py
@@ -5,10 +5,12 @@ from typing import List
 
 import tiledb
 
+from backend.wmg.data.schemas.corpus_schema import OBS_ARRAY_NAME
+
 
 @cache
 def get_all_dataset_ids(tdb_group: str) -> List[str]:
-    with tiledb.open(f"{tdb_group}/obs", "r") as obs:
+    with tiledb.open(f"{tdb_group}/{OBS_ARRAY_NAME}", "r") as obs:
         all_dataset_ids = obs.query(attrs=[], dims=["dataset_id"]).df[:].dataset_id.unique()
     all_dataset_ids.sort()
     return all_dataset_ids
@@ -25,3 +27,10 @@ def log_func_runtime(func):
         return result
 
     return wrap_func
+
+
+def create_empty_cube(uri: str, schema):
+    """
+    Create an empty cube with expected schema (dimensions and attributes) at given uri
+    """
+    tiledb.Array.create(uri, schema, overwrite=True)
